backend/requirements-test.txt
# Testing Dependencies for SafeSphere

# Core Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0  # Parallel test execution
pytest-mock==3.12.0

# Test Data Generation
factory-boy==3.3.0
faker==22.0.0

# Time Mocking
freezegun==1.4.0

# Load Testing
locust==2.20.0

# Performance Profiling
django-silk==5.0.4
django-debug-toolbar==4.2.0

# Security Testing
bandit==1.7.6
safety==3.0.1

# Code Quality
flake8==7.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0

# API Testing
requests==2.31.0
responses==0.24.1

# Database Testing
pytest-randomly==3.15.0  # Randomize test order
pytest-timeout==2.2.0    # Timeout long-running tests

# Coverage Reporting
coverage[toml]==7.4.1

